+++ order/src/main/java/com/toanyone/order/application/service/StoreService.java
@@ -1,7 +1,7 @@
-package com.toanyone.order.application;
+package com.toanyone.order.application.service;
 
 import com.toanyone.order.application.dto.StoreFindResponseDto;
-import com.toanyone.order.common.SingleResponse;
+import com.toanyone.order.common.dto.SingleResponse;
 import org.springframework.http.ResponseEntity;
 
 public interface StoreService {
