+++ order/src/main/java/com/toanyone/order/common/dto/CursorPage.java
@@ -1,4 +1,4 @@
-package com.toanyone.order.common;
+package com.toanyone.order.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
