+++ order/src/main/java/com/toanyone/order/common/dto/SingleResponse.java
@@ -1,4 +1,4 @@
-package com.toanyone.order.common;
+package com.toanyone.order.common.dto;
 
 import lombok.Getter;
