+++ order/src/main/java/com/toanyone/order/domain/repository/OrderRepository.java
@@ -1,10 +1,9 @@
 package com.toanyone.order.domain.repository;
 
 import com.toanyone.order.application.dto.request.OrderSearchCondition;
-import com.toanyone.order.common.CursorPage;
-import com.toanyone.order.domain.entity.Order;
-import com.toanyone.order.infrastructure.OrderQueryDslRepository;
-import com.toanyone.order.presentation.dto.request.OrderSearchRequestDto;
+import com.toanyone.order.common.dto.CursorPage;
+import com.toanyone.order.domain.model.Order;
+import com.toanyone.order.infrastructure.repository.OrderQueryDslRepository;
 
 import java.util.Optional;
