+++ order/src/main/java/com/toanyone/order/infrastructure/client/ItemClient.java
@@ -1,10 +1,9 @@
 package com.toanyone.order.infrastructure.client;
 
-import com.toanyone.order.application.ItemService;
+import com.toanyone.order.application.service.ItemService;
 import com.toanyone.order.application.dto.ItemRestoreRequestDto;
 import com.toanyone.order.application.dto.ItemValidationRequestDto;
 import jakarta.validation.Valid;
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
