+++ order/src/main/java/com/toanyone/order/presentation/dto/request/OrderFindAllRequestDto.java
@@ -1,13 +1,13 @@
 package com.toanyone.order.presentation.dto.request;
 
-import com.toanyone.order.common.SortType;
+import com.toanyone.order.common.dto.SortType;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
-import static com.toanyone.order.common.CursorConstants.DEFAULT_SIZE;
-import static com.toanyone.order.common.CursorConstants.DEFAULT_SORT_TYPE;
+import static com.toanyone.order.common.dto.CursorConstants.DEFAULT_SIZE;
+import static com.toanyone.order.common.dto.CursorConstants.DEFAULT_SORT_TYPE;
 
 @Getter
 public class OrderFindAllRequestDto {
