+++ order/src/main/java/com/toanyone/order/presentation/dto/request/OrderSearchRequestDto.java
@@ -1,12 +1,12 @@
 package com.toanyone.order.presentation.dto.request;
 
-import com.toanyone.order.common.SortType;
+import com.toanyone.order.common.dto.SortType;
 import lombok.*;
 
 import java.time.LocalDateTime;
 
-import static com.toanyone.order.common.CursorConstants.DEFAULT_SIZE;
-import static com.toanyone.order.common.CursorConstants.DEFAULT_SORT_TYPE;
+import static com.toanyone.order.common.dto.CursorConstants.DEFAULT_SIZE;
+import static com.toanyone.order.common.dto.CursorConstants.DEFAULT_SORT_TYPE;
 
 @Getter
 public class OrderSearchRequestDto {
