+++ order/src/main/java/com/toanyone/order/presentation/dto/response/OrderCreateResponseDto.java
@@ -1,7 +1,7 @@
 package com.toanyone.order.presentation.dto.response;
 
-import com.toanyone.order.domain.entity.Order;
-import com.toanyone.order.domain.entity.OrderItem;
+import com.toanyone.order.domain.model.Order;
+import com.toanyone.order.domain.model.OrderItem;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
